fix more complaints

+++ server/src/net/respond.c
@@ -25,7 +25,7 @@ void net_respond(struct server *s)
 			break;
 		case server_message_chunk:
 			b += pack_sm_chunk(sm->update, &buf[b]);
-			L("packed up a chunk, %d bytes", b);
+			L("packed up a chunk, %ld bytes", (long)b);
 			break;
 		}
